# src/main/java/com/WebJava/spacecatsmarket/dto/customer/CustomerDetailsListDto.java
package com.WebJava.spacecatsmarket.dto.customer;

import java.util.List;
import lombok.Builder;
import lombok.Value;
import lombok.extern.jackson.Jacksonized;

@Value
@Builder(toBuilder = true)
@Jacksonized
public class CustomerDetailsListDto {

    List<CustomerDetailsEntry> customerDetailsEntries;

}