# src/main/java/com/WebJava/spacecatsmarket/web/exception/ParamsViolationDetails.java
package com.WebJava.spacecatsmarket.web.exception;

import lombok.Builder;
import lombok.Value;
import lombok.extern.jackson.Jacksonized;

@Value
@Builder(toBuilder = true)
@Jacksonized
public class ParamsViolationDetails {
    String fieldName;
    String reason;
}